chore: move timer to the logging module
